+++ src/DistanceTabGenerator.java
@@ -4,10 +4,10 @@ public class DistanceTabGenerator {
 	
 	private static int _nbCall = 0;
 	
-	public static int[][] generateTab(int nbLigne, int minValue, int maxValue) {
+	public static double[][] generateTab(int nbLigne, int minValue, int maxValue) {
 		_nbCall++;
 		Random r = new Random(_nbCall);
-		int[][] distanceTab = new int[nbLigne][nbLigne];
+		double[][] distanceTab = new double[nbLigne][nbLigne];
 		for (int i =0 ; i< nbLigne;i++){
 			for (int j = i+1 ; j <nbLigne;j++){
 				distanceTab [i][j]=r.nextInt(maxValue-minValue+1)+minValue;
