+++ src/FloydWarshallSolver.java
@@ -1,13 +1,13 @@
 
 public class FloydWarshallSolver {
-	 public static void reduce(int[][] distanceR) {
+	 public static void reduce(double[][] distanceR) {
 	       int lastIndexChange = 0;
 	       int start = 0;
 	       int nbNode = distanceR.length;
 	       do {
 	           for (int end = 0; end < nbNode; end++) {
 	               for (int intermediate = 0; intermediate < nbNode; intermediate++) {
-	                   int newValue = distanceR[start][intermediate] + distanceR[intermediate][end];
+	                   double newValue = distanceR[start][intermediate] + distanceR[intermediate][end];
 	                   if (distanceR[start][end] > newValue) {
 	                       distanceR[start][end] = newValue;
 	                       distanceR[end][start] = newValue;
